Refactor (legacy): move custom GUI components to current code. [Not yet in use.]
