--- legacy/probably_useful/source/util/AbortException.java
@@ -1,19 +1,0 @@
-/**
- * 
- */
-package ch.unibas.landolt.balduin.XMLSpeedupV6.src.util;
-
-/**
- * @author Balduin Landolt
- *
- */
-public class AbortException extends Exception {
-
-	/**
-	 * @param string
-	 */
-	public AbortException(String msg) {
-		super(msg);
-	}
-	
-}
