--- legacy/probably_useful/source/util/Loggable.java
@@ -1,12 +1,0 @@
-/**
- * 
- */
-package ch.unibas.landolt.balduin.XMLSpeedupV6.src.util;
-
-/**
- * @author Balduin Landolt
- *
- */
-public interface Loggable {
-	public void log(String s);
-}
