+++ legacy/probably_useless/ActionCommands.java
@@ -1,0 +1,27 @@
+/**
+ * 
+ */
+package ch.unibas.landolt.balduin.XMLSpeedupV6.src.util;
+
+/**
+ * @author Balduin Landolt
+ *
+ */
+public class ActionCommands {
+	public static final String newTranscription = "newTranscription";
+	public static final String importXML = "importXML";
+	public static final String importRaw = "importRaw";
+	public static final String save = "save";
+	public static final String saveAs = "saveAs";
+	public static final String close = "close";
+	public static final String openFolder = "openFolder";
+	public static final String undo = "undo";
+	public static final String redo = "redo";
+	public static final String loadImages = "loadImages";
+	public static final String lineEntered = "lineEntered";
+	public static final String previewContentChanged = "previewContentChanged";
+	public static final String zoomIn = "zoomIn";
+	public static final String zoomOut = "zoomOut";
+	public static final String cropSeletedArea = "cropSelectedArea";
+	public static final String CtrlTabPressed = "CtrlTabPressed";
+}
