+++ src/main/java/ch/blandolt/turboTranscriber/util/AbortException.java
@@ -1,0 +1,19 @@
+/**
+ * 
+ */
+package ch.blandolt.turboTranscriber.util;
+
+/**
+ * @author Balduin Landolt
+ *
+ */
+public class AbortException extends Exception {
+
+	/**
+	 * @param string
+	 */
+	public AbortException(String msg) {
+		super(msg);
+	}
+	
+}
